Divide business logic into separate files
