+++ tests/unit/test_pool.py
@@ -1,0 +1,47 @@
+from iwant_bot import requests
+from iwant_bot.storage import MemoryRequestsStorage
+from iwant_bot.pool import RequestsPool
+
+
+def make_filled_request_pool():
+    storage = MemoryRequestsStorage()
+    storage.store_request(requests.IWantRequest("john", "coffee", 5))
+    storage.store_request(requests.IWantRequest("jane", "coffee", 5))
+    storage.store_request(requests.IWantRequest("olivia", "picnic", 5))
+    storage.store_request(requests.IWantRequest("jerry", "coffee", 0))
+    pool = RequestsPool(storage)
+    pool.update_requests_from_storage()
+    return pool
+
+
+def test_requests_pool_filters_relevant():
+    pool = make_filled_request_pool()
+    active_requests = pool.current_activities_requests
+    assert len(active_requests) == 3
+
+
+def test_request_cancel():
+    pool = make_filled_request_pool()
+    storage = pool._requests_storage
+
+    cancel_picnic = requests.ICancelRequest("olivia", "picnic", 2)
+    storage.store_request(cancel_picnic)
+    pool.update_requests_from_storage()
+    active_requests = pool.current_activities_requests
+    assert len(active_requests) == 2
+
+    cancel_coffee_late = requests.ICancelRequest("john", "coffee", 6)
+    storage.store_request(cancel_coffee_late)
+    pool.update_requests_from_storage()
+    active_requests = pool.current_activities_requests
+    assert len(active_requests) == 2
+
+
+def test_pool_ignores_duplicates():
+    pool = make_filled_request_pool()
+    storage = pool._requests_storage
+    storage.store_request(requests.IWantRequest("john", "coffee", 5))
+    storage.store_request(requests.IWantRequest("james", "coffee", 5))
+    pool.update_requests_from_storage()
+    active_requests = pool.current_activities_requests
+    assert len(active_requests) == 4
